requirements.txt
langchain==0.3.15
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-groq==0.2.1
langchain-text-splitters==0.3.4

openai>=1.6.0
tiktoken>=0.7.0

faiss-cpu==1.9.0
pypdf==5.1.0
streamlit==1.40.2
python-dotenv==1.0.1

ragas==0.2.9
datasets==3.2.0
pandas==2.2.3
numpy>=1.26.0

redis==5.2.1
tenacity>=8.2.3

duckduckgo-search==6.3.5
tqdm>=4.66.0
